Fix issue with hashing_function that was causing it to hash integers improperly
The hashing_function calls the [] method on whatever is passed into the hasing_function. I assumed that anything passed into the hashing_function would be a string...so this made sense. However, I was also passing in *integers* into the hashing_function. Turns out that integers can also respond to the [] method...

"Returns the nth bit in the binary representation of fix, where fix[0] is the least significant bit. Returns: (0,1)"

See Docs - http://www.rubydoc.info/stdlib/core/Fixnum:[]

I was relying on the hashing_function to provide me unique values, but since the only values I'd get are 0 and 1, I have much more hashing collisions.

The issue with the hashing_function was responsible for 30% of tests failing. Fixing the issue ensure that all tests pass. However, the real problem is that the hashing_function is just not "good", period. I need a better hashing_function in the long-term.

DUCK-TYPING.