Handle hashing collisions by storing key-value pairs by using a Linked List
When we access an index, we then tranverse the linked list looking for the node that contains the key that we're looking for, and once we find it, we output the value. Most of the time, we will never have to worry about looking past the first node, so we still have an O(1) lookup time. To keep #remove at O(1), we "pretend" to delete the node, but really just setting the value to nil.

This commit winds up breaking HashSet, which was heavily dependent on #size. I'm going to fix this by removing #size from MyHash and creating my own custom method for HashSet.