Remove #size method
The only reason this method existed was to act as a helper method for HashSet, but this helper method seems to fail when we changed the implementation of MyHash to use Linked Lists in addition to array size. Since we deleted #size, we're also deleting the test associated with #size.