Remove #size memoization in favor of simply delegating #size to the element_array of HashSet
I forgot that I was keeping a list of all the keys of a HashSet already so that we can iteriate over the HashSet later on. If we want to kow the size of a HashSet, we can simply take a look at the list of keys in the HashSet.